requirements.txt
Flask==2.3.2
flask-cors==4.0.0
Flask-SQLAlchemy==3.1.1
pymysql
python-dotenv